refactor: Refactored containerVariants and cardVariants into own files
