+++ src/app/gallery/_components/GallaryImages.tsx
@@ -1,6 +1,8 @@
 import { FC } from 'react';
 import Image from 'next/image';
 import { motion } from 'motion/react';
+import containerVariants from '@/styles/containerVariants';
+
 
 interface ImagesProps {
   data: {
@@ -15,21 +17,6 @@ const GalleryImages: FC<ImagesProps> = (props) => {
 
   const handleClick = (index: number) => {
     onClick(index);
-  };
-
-  const containerVariants = {
-    hidden: { opacity: 0 },
-    visible: {
-      opacity: 1,
-      transition: {
-        staggerChildren: 0.1,
-      },
-    },
-  };
-
-  const cardVariants = {
-    hidden: { opacity: 0, y: 50 },
-    visible: { opacity: 1, y: 0, transition: { duration: 0.5 } },
   };
 
   return (
