+++ src/styles/cardVariants.ts
@@ -1,0 +1,6 @@
+const cardVariants = {
+  hidden: { opacity: 0, y: 50 },
+  visible: { opacity: 1, y: 0, transition: { duration: 0.5 } },
+};
+
+export default cardVariants
