+++ src/styles/containerVariants.ts
@@ -1,0 +1,11 @@
+const containerVariants = {
+  hidden: { opacity: 0 },
+  visible: {
+    opacity: 1,
+    transition: {
+      staggerChildren: 0.1,
+    },
+  },
+};
+
+export default containerVariants;
